Use mixin pattern for abstract analyzer.
Due to missing the power of super I thought this was not possible.
Thanks to a better understanding due to the kind explanation by @terrencepreilly
in 382e2de8a56a544f62ced97af261b76d00b0808a this is now corrected.
