+++ darglint/analysis/abstract_callable_visitor.py
@@ -4,25 +4,119 @@ from typing import (
     Dict,
     List,
 )
-from .pure_abstract_analyzer import (
-    analyze_pure_abstract,
-)
 
 
 class AbstractCallableVisitor(ast.NodeVisitor):
 
     def __init__(self, *args, **kwargs):
         # type: (List[Any], Dict[str, Any]) -> None
-        super(AbstractCallableVisitor, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
         self.is_abstract = None
 
+    def _is_docstring(self, node):
+        # type: (ast.Return) -> bool
+        return (
+            isinstance(node, ast.Expr) and (
+                (
+                    isinstance(node.value, ast.Constant) and
+                    isinstance(node.value.value, str)
+                ) or (
+                    isinstance(node.value, ast.Str) # Python < 3.8
+                )
+            )
+        )
+
+    def _is_ellipsis(self, node):
+        # type: (ast.Return) -> bool
+        print(f"{getattr(node,'value','')} ellipsis")
+
+        return (
+            isinstance(node, ast.Expr) and (
+                (
+                    isinstance(node.value, ast.Constant) and
+                    node.value.value is Ellipsis
+                ) or (
+                    isinstance(node.value, ast.Ellipsis) # Python < 3.8
+                )
+            )
+        )
+
+    def _is_raise_NotImplementedException(self, node):
+        # type: (ast.Return) -> bool
+        return (
+            isinstance(node, ast.Raise) and (
+                (
+                    isinstance(node.exc, ast.Name) and
+                    node.exc.id == "NotImplementedError"
+                ) or (
+                    isinstance(node.exc, ast.Call)
+                    and isinstance(node.exc.func, ast.Name)
+                    and node.exc.func.id == "NotImplementedError"
+                )
+            )
+        )
+
+    def _is_return_NotImplemented(self, node):
+        # type: (ast.Return) -> bool
+        return (
+            isinstance(node, ast.Return) and
+            isinstance(node.value, ast.Name) and
+            node.value.id == "NotImplemented"
+        )
+
+    def _is_abstract(self, node):
+        # type: (ast.Return) -> bool
+        for decorator in node.decorator_list:
+            # apparently there can be decorators without id's
+            if getattr(decorator, "id", "") == ("abstractmethod"):
+                return True
+
+        return False
+
+    def analyze_pure_abstract(self, node):
+        # type: (ast.Return) -> bool
+        assert isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef)), (
+            "Assuming this analysis is only called on functions"
+        )
+
+        if not self._is_abstract(node):
+            return False
+
+        children = len(node.body)
+
+        # maximum docstring and one statement
+        if children > 2:
+            return False
+
+        if children == 2:
+            if not self._is_docstring(node.body[0]):
+                return False
+
+            statement = node.body[1]
+        else:
+            statement = node.body[0]
+
+        if (
+            isinstance(statement, ast.Pass) or
+            self._is_ellipsis(statement) or
+            self._is_raise_NotImplementedException(statement) or
+            self._is_return_NotImplemented(statement) or
+            (
+                children == 1 and
+                self._is_docstring(statement)
+            )
+        ):
+            return True
+
+        return False
+
     def visit_FunctionDef(self, node):
         # type: (ast.FunctionDef) -> ast.AST
-        self.is_abstract = analyze_pure_abstract(node)
-        return super().visit_FunctionDef(node)
+        self.is_abstract = self.analyze_pure_abstract(node)
+        return getattr(super(), "visit_FunctionDef", super().generic_visit)(node)
 
     def visit_AsyncFunctionDef(self, node):
         # type: (ast.AsyncFunctionDef) -> ast.AST
-        self.is_abstract = analyze_pure_abstract(node)
-        return super().visit_AsyncFunctionDef(node)
+        self.is_abstract = self.analyze_pure_abstract(node)
+        return getattr(super(), "visit_AsyncFunctionDef", super().generic_visit)(node)
